+++ operators/advanced_network/advanced_network/common.cpp
@@ -344,30 +344,36 @@ Status adv_net_init(NetworkConfig& config) {
 // RDMA Functions
 Status rdma_connect_to_server(const std::string& server_addr, uint16_t server_port,
                               uintptr_t* conn_id) {
+  ASSERT_ANO_MGR_INITIALIZED();
   return g_ano_mgr->rdma_connect_to_server(server_addr, server_port, conn_id);
 }
 
 Status rdma_connect_to_server(const std::string& server_addr, uint16_t server_port,
                               const std::string& src_addr, uintptr_t* conn_id) {
+  ASSERT_ANO_MGR_INITIALIZED();
   return g_ano_mgr->rdma_connect_to_server(server_addr, server_port, src_addr, conn_id);
 }
 
 Status rdma_get_port_queue(uintptr_t conn_id, uint16_t* port, uint16_t* queue) {
+  ASSERT_ANO_MGR_INITIALIZED();
   return g_ano_mgr->rdma_get_port_queue(conn_id, port, queue);
 }
 
 Status rdma_get_server_conn_id(const std::string& server_addr, uint16_t server_port,
                                uintptr_t* conn_id) {
+  ASSERT_ANO_MGR_INITIALIZED();
   return g_ano_mgr->rdma_get_server_conn_id(server_addr, server_port, conn_id);
 }
 
 Status rdma_set_header(BurstParams* burst, RDMAOpCode op_code, uintptr_t conn_id, bool is_server,
                        int num_pkts, uint64_t wr_id, const std::string& local_mr_name) {
+  ASSERT_ANO_MGR_INITIALIZED();
   return g_ano_mgr->rdma_set_header(
       burst, op_code, conn_id, is_server, num_pkts, wr_id, local_mr_name);
 }
 
 RDMAOpCode rdma_get_opcode(BurstParams* burst) {
+  ASSERT_ANO_MGR_INITIALIZED();
   return g_ano_mgr->rdma_get_opcode(burst);
 }