+++ operators/advanced_network/CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Non-Breaking
 
--   **RDMA Backend**
-  - Added RDMA backend supporting both client/server with RDMA-CM. SEND/RECV supported currently
+RDMA Backend:
+-   **New RDMA backend**
+    - Added RDMA backend supporting both client/server with RDMA-CM. SEND/RECV supported currently
 
 ## 🚧 holoscan-networking 0.2